fix: correct  add and get user

+++ src/database/UsersDb.ts
@@ -1,5 +1,5 @@
 import { v4 as uuidv4 } from 'uuid';
-import { IServerResolve } from '../types/types';
+import { IServerResolve, IServerAnswer } from '../types/types';
 
 interface IUserData {
   username: string;
@@ -14,8 +14,14 @@ class UsersDb {
     this.database = new Map();
   }
 
-  addUser(data: IUserData) {
+  addUser(data: IUserData): IServerAnswer {
     const uniqueId = uuidv4();
+    //ToDo check data.hobbies on empty array
+    // if (!data.age || !data.hobbies || !data.username) {
+    //   return {
+    //     status: { statusCode: 404, message: 'request body does not contain required fields' },
+    //   };
+    // }
     const userData: IServerResolve = {
       id: uniqueId,
       username: data.username,
@@ -24,24 +30,29 @@ class UsersDb {
     };
 
     this.database.set(uniqueId, userData);
+    return { status: { statusCode: 201, message: 'create new user' } };
+  }
+  //ToDo return all users
+  getUsers(): IServerAnswer {
+    // console.log('map.values');
+    // console.log(this.database.values());
+    // console.log('map.entries');
+    // console.log(this.database.entries());
+    const userArr = Array.from(this.database.values());
+    return { resolve: userArr };
   }
 
-  getUsers() {
-    console.log('map.values');
-    console.log(this.database.values());
-    console.log('map.entries');
-    console.log(this.database.entries());
-  }
-
-  getUsersById(id: string): IServerResolve {
-    return this.database.get(id);
+  getUsersById(id: string): IServerAnswer {
+    return this.isIdExist(id)
+      ? { status: { statusCode: 200, message: `id === ${id}` } }
+      : { status: { statusCode: 404, message: 'id is not exist' } };
   }
   //ToDo check is id UUIDV4
   isIdExist(id: string): boolean {
     return this.database.has(id);
   }
 
-  updateUserById(id: string, data: IUserData): void {
+  updateUserById(id: string, data: IUserData): IServerAnswer {
     if (this.isIdExist(id)) {
       const userData: IServerResolve = {
         id: id,
@@ -51,11 +62,17 @@ class UsersDb {
       };
 
       this.database.set(id, userData);
+      return { status: { statusCode: 200, message: 'user info was update' } };
     }
+    return { status: { statusCode: 404, message: 'id is not exist' } };
   }
 
-  deleteUserById(id: string): void {
-    this.database.delete(id);
+  deleteUserById(id: string): IServerAnswer {
+    if (this.isIdExist(id)) {
+      this.database.delete(id);
+      return { status: { statusCode: 204, message: 'user info was deleted' } };
+    }
+    return { status: { statusCode: 404, message: 'id is not exist' } };
   }
 }
